pyproject.toml
[project]
name = "rutube-downloader"
version = "0.1.0"
description = "Скрипт для скачивания видео с Rutube"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "m3u8>=3.5.0",
    "tqdm>=4.66.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

